requirements.txt
clip @ git+https://github.com/openai/CLIP.git@dcba3cb2e2827b402d2701e7e1c7d9fed8a20ef1
ftfy==6.2.3
huggingface-hub==0.35.3
matplotlib==3.7.5
numpy==1.24.4
pandas==2.0.3
pillow==10.4.0
POT==0.9.0
scikit-learn==1.3.2
scipy==1.10.1
timm==1.0.20
tokenizers==0.20.3
torch==2.3.1 --index-url https://download.pytorch.org/whl/cu121
torchvision==0.18.1 --index-url https://download.pytorch.org/whl/cu121
torchaudio==2.3.1 --index-url https://download.pytorch.org/whl/cu121
tqdm==4.67.1
transformers==4.46.3

